Moved HIV diagnosis to its own module and redistributed some code.
